requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
Flask==2.3.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
fonttools==4.44.0
gevent==23.9.1
greenlet==3.0.1
html5lib==1.1
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
memory-profiler==0.61.0
Pillow==10.1.0
prompt-toolkit==3.0.39
psutil==5.9.6
pycparser==2.21
pydyf==0.8.0
PyJWT==2.8.0
pyparsing==3.1.1
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
weasyprint==60.1
webencodings==0.5.1
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
zopfli==0.2.3
